Introduced module containing all class modifictions required for testing
have all class_eval statements at one common place. reduced unknown side
effects. applying class_eval in different tests on the same target class
results in having unknown side effects.
